working with shiko

+++ OS_Project/process.h
@@ -47,7 +47,7 @@ private:
     qreal sumBursttime();
     int processTakePriority(qreal time, unsigned int currentProcess);
     int processTakeSJF(qreal time, qreal gnttchart,unsigned int currentProcess);
-    void RR_queueProcessing(QQueue <unsigned int> &, unsigned int , unsigned int &, QList <qreal> &);
+    void RR_queueProcessing(QQueue <unsigned int> &, unsigned int , qreal &, QList <qreal> &);
 
 
 public:
